remove combinedReducer to use smurfReducer only
